HIRES pipeline: minor updates
