+++ gamse/pipelines/hires/common.py
@@ -1,5 +1,18 @@
-import numpy as np
+import os
+import re
+import logging
+logger = logging.getLogger(__name__)
 
+import numpy as np
+import astropy.io.fits as fits
+from scipy.ndimage.filters import gaussian_filter
+from scipy.signal import savgol_filter
+import scipy.interpolate as intp
+import matplotlib.pyplot as plt
+import matplotlib.ticker as tck
+from matplotlib.figure import Figure
+
+from ...echelle.imageproc import combine_images
 from ...echelle.trace import TraceFigureCommon
 
 def print_wrapper(string, item):
@@ -169,7 +182,6 @@ def get_badpixel_mask(binning, ccd=0):
             mask[:,         0:45] = True
     return np.int16(mask)
 
-
 def mosaic_3_images(data_lst, mask_lst):
     """Mosaic three images.
 
@@ -207,11 +219,324 @@ def mosaic_3_images(data_lst, mask_lst):
     return allimage, allmask
 
 
+def get_bias_post2004(config, logtable):
+    """Get bias image.
+
+    Args:
+        config (:class:`configparser.ConfigParser`): Config object.
+        logtable (:class:`astropy.table.Table`): Table of Observing log.
+
+    Returns:
+        tuple: A tuple containing:
+
+            * **bias_lst** (list) – A list of :class:`numpy.ndarray` as the
+              output bias image for 3 CCDs.
+            * **bias_card_lst** (list) – List of FITS header cards related to
+              the bias correction.
+    """
+
+    nccd = 3
+    mode = config['reduce'].get('mode')
+    bias_file = config['reduce.bias'].get('bias_file')
+
+    if mode=='debug' and os.path.exists(bias_file):
+        # load bias data from existing file
+        hdu_lst = fits.open(bias_file)
+        # pack bias image
+        bias_lst = [hdu_lst[iccd+1+nccd].data for iccd in range(nccd)]
+        hdu_lst.close()
+
+        reobj = re.compile('GAMSE BIAS[\s\S]*')
+        # filter header cards that match the above pattern
+        bias_card_lst = [(card.keyword, card.value) for card in head.cards
+                            if reobj.match(card.keyword)]
+
+        message = 'Load bias data from file: "{}"'.format(bias_file)
+        logger.info(message)
+        print(message)
+    else:
+        bias_lst, bias_card_lst = combine_bias_post2004(config, logtable)
+
+    return bias_lst, bias_card_lst
+
+def combine_bias_post2004(config, logtable):
+    """Combine the bias images.
+
+    Args:
+        config (:class:`configparser.ConfigParser`): Config object.
+        logtable (:class:`astropy.table.Table`): Table of Observing log.
+
+    Returns:
+        tuple: A tuple containing:
+
+            * **bias** (:class:`numpy.ndarray`) – Output bias image.
+            * **bias_card_lst** (list) – List of FITS header cards related to
+              the bias correction.
+
+    """
+    nccd = 3
+
+    rawpath = config['data']['rawpath']
+
+    # determine number of cores to be used
+    ncores = config['reduce'].get('ncores')
+    if ncores == 'max':
+        ncores = os.cpu_count()
+    else:
+        ncores = min(os.cpu_count(), int(ncores))
+
+    section = config['reduce.bias']
+    bias_file = section['bias_file']
+
+    bias_data_lst = [[] for iccd in range(nccd)]
+    bias_card_lst = []
+
+    bias_items = list(filter(lambda item: item['object'].lower()=='bias',
+                             logtable))
+    # get the number of bias images
+    n_bias = len(bias_items)
+
+    if n_bias == 0:
+        # there is no bias frames
+        return None, []
+
+    fmt_str = '  - {:>7s} {:^17} {:^20s} {:^7}'
+    head_str = fmt_str.format('frameid', 'FileID', 'Object', 'exptime')
+
+    for iframe, logitem in enumerate(bias_items):
+
+        fname = '{}.fits'.format(logitem['fileid'])
+        filename = os.path.join(rawpath, fname)
+        hdu_lst = fits.open(filename)
+        data_lst, mask_lst = parse_3ccd_images(hdu_lst)
+        hdu_lst.close()
+
+        for iccd in range(nccd):
+            bias_data_lst[iccd].append(data_lst[iccd])
+
+        # print info
+        if iframe == 0:
+            print('* Combine Bias Images: {}'.format(bias_file))
+            print(head_str)
+        message = fmt_str.format(
+                    '[{:d}]'.format(logitem['frameid']),
+                    logitem['fileid'], logitem['object'],
+                    logitem['exptime']
+                    )
+        print(message)
+
+    prefix = 'HIERARCH GAMSE BIAS '
+    bias_card_lst.append((prefix + 'NFILE', n_bias))
+
+    combine_mode = 'mean'
+    section = config['reduce.bias']
+    cosmic_clip  = section.getfloat('cosmic_clip')
+    maxiter      = section.getint('maxiter')
+    maskmode    = (None, 'max')[n_bias>=3]
+
+    bias_card_lst.append((prefix+'COMBINE_MODE', combine_mode))
+    bias_card_lst.append((prefix+'COSMIC_CLIP',  cosmic_clip))
+    bias_card_lst.append((prefix+'MAXITER',      maxiter))
+    bias_card_lst.append((prefix+'MASK_MODE',    str(maskmode)))
+
+    # create the hdu list to be saved
+    hdu_lst = fits.HDUList()
+    # create new FITS Header for bias
+    head = fits.Header()
+    for card in bias_card_lst:
+        head.append(card)
+    head['HIERARCH GAMSE FILECONTENT 0'] = 'NONE'
+
+    # pack the primary HDU
+    hdu_lst.append(fits.PrimaryHDU(header=head))
+
+    # prepare the list for image data of each CCD
+    bias_lst = []
+
+    # scan for each ccd
+    for iccd in range(nccd):
+        ### 3 CCDs loop begins here ###
+        bias_data_lst[iccd] = np.array(bias_data_lst[iccd])
+
+        sub_bias = combine_images(bias_data_lst[iccd],
+            mode        = combine_mode,
+            upper_clip  = cosmic_clip,
+            maxiter     = maxiter,
+            maskmode    = maskmode,
+            ncores      = ncores,
+            )
+
+        # pack bias of each CCD into sub_bias_lst
+        bias_lst.append(sub_bias)
+
+        message = ('\033[{}mCombined bias for CCD {}: '
+                    'Mean = {:6.2f}\033[0m'.format(
+                    (34, 32, 31)[iccd], iccd+1, sub_bias.mean()))
+        print(message)
+
+        key = 'HIERARCH GAMSE FILECONTENT {}'.format(iccd+1)
+        hdu_lst[0].header[key] = 'BIAS COMBINED FOR CCD{}'.format(iccd+1)
+        hdu_lst.append(fits.ImageHDU(data=sub_bias))
+
+    # initialize bias figure
+    bias_fig = BiasFigure(data_lst=bias_lst)
+
+    new_bias_lst = []
+
+    # calculate and plot ymean of bias data
+    for iccd in range(nccd):
+        data = bias_lst[iccd]
+        ny, nx = data.shape
+        ally = np.arange(ny)
+        ymean = data.mean(axis=1)
+        ax_ycut = bias_fig.ax_ycut_lst[iccd]
+        ax_ycut.plot(ymean, ally, color='C3', lw=0.5, alpha=0.6)
+
+        mask = np.ones_like(ymean, dtype=np.bool)
+        for i in range(10):
+            f = intp.InterpolatedUnivariateSpline(
+                    ally[mask], ymean[mask], k=1)
+            yrec = f(ally)
+            ysmo = savgol_filter(yrec, window_length=301, polyorder=3)
+            yres = ymean - ysmo
+            ystd = yres[mask].std()
+            newmask = (yres>-3*ystd)*(yres<3*ystd)
+            if newmask.sum()==mask.sum():
+                break
+            mask = newmask
+
+        ax_ycut.plot(ysmo, ally, color='r', lw=0.5, alpha=1)
+
+        new_bias = np.tile(ysmo, (nx,1)).T
+        new_bias_lst.append(new_bias)
+
+    # save and close the figure
+    figpath = config['reduce']['figpath']
+    figfilename = os.path.join(figpath, 'bias.png')
+    bias_fig.savefig(figfilename)
+    plt.close(bias_fig)
+
+    for iccd in range(nccd):
+        hdu_lst.append(fits.ImageHDU(data=new_bias_lst[iccd]))
+        # update the file content in primary HDU
+        key = 'HIERARCH GAMSE FILECONTENT {}'.format(1+nccd+iccd)
+        card = (key, 'BIAS USED FOR CCD{}'.format(iccd+1))
+        hdu_lst[0].header.append(card)
+    bias_lst = new_bias_lst
+
+
+    ################ bias smooth ################
+    if section.getboolean('smooth'):
+        # bias needs to be smoothed
+        smooth_method = section.get('smooth_method')
+
+        newcard_lst = []
+        if smooth_method in ['gauss', 'gaussian']:
+            # perform 2D gaussian smoothing
+            smooth_sigma = section.getint('smooth_sigma')
+            smooth_mode  = section.get('smooth_mode')
+            
+            for iccd in range(nccd):
+                sub_bias = bias_lst[iccd]
+                bias_smooth = gaussian_filter(sub_bias,
+                                sigma = smooth_sigma,
+                                mode  = smooth_mode,
+                                )
+                hdu_lst.append(fits.ImageHDU(data=bias_smooth))
+
+                # update the file content in primary HDU
+                key = 'HIERARCH GAMSE FILECONTENT {}'.format(1+nccd+iccd)
+                card = (key, 'BIAS SMOOTHED FOR CCD{}'.format(iccd+1))
+                hdu_lst[0].header.append(card)
+
+                # udpate the result array to be returned
+                bias_lst[iccd] = bias_smooth
+
+            # write information to FITS header
+            newcard_lst.append((prefix + 'SMOOTH CORRECTED', True))
+            newcard_lst.append((prefix + 'SMOOTH METHOD',  'GAUSSIAN'))
+            newcard_lst.append((prefix + 'SMOOTH SIGMA',   smooth_sigma))
+            newcard_lst.append((prefix + 'SMOOTH MODE',    smooth_mode))
+        else:
+            print('Unknown smooth method: ', smooth_method)
+            pass
+
+        # pack the cards to bias_card_lst and also hdu_lst
+        for card in newcard_lst:
+            hdu_lst[0].header.append(card)
+            bias_card_lst.append(card)
+
+    ############### save to FITS ##############
+    hdu_lst.writeto(bias_file, overwrite=True)
+
+    message = 'Bias image written to "{}"'.format(bias_file)
+    logger.info(message)
+    print(message)
+
+    return bias_lst, bias_card_lst
+
+class BiasFigure(Figure):
+
+    def __init__(self, dpi=200, figsize=(12,8), data_lst=None):
+        Figure.__init__(self, dpi=dpi, figsize=figsize)
+        nccd = 3
+        axh = 0.8
+
+        self.ax_imag_lst = []
+        self.ax_ycut_lst = []
+        self.ax_hist_lst = []
+        self.ax_cbar_lst = []
+        for iccd in range(nccd):
+            data = data_lst[iccd]
+            ny, nx = data.shape
+            axw = axh/figsize[0]*figsize[1]/ny*nx
+            ax_imag = self.add_axes([0.1+iccd*0.20, 0.07, axw, axh])
+            ax_ycut = ax_imag.twiny()
+            ax_hist = self.add_axes([0.7, 0.07+(nccd-iccd-1)*0.28+0.03, 0.25, 0.2])
+            ax_cbar = self.add_axes([0.7, 0.07+(nccd-iccd-1)*0.28, 0.25, 0.02])
+
+            self.ax_imag_lst.append(ax_imag)
+            self.ax_ycut_lst.append(ax_ycut)
+            self.ax_hist_lst.append(ax_hist)
+            self.ax_cbar_lst.append(ax_cbar)
+
+            vmin = np.percentile(data, 1)
+            vmax = np.percentile(data, 99)
+            cax = ax_imag.imshow(data, origin='lower', vmin=vmin, vmax=vmax)
+            self.colorbar(cax, cax=ax_cbar, orientation='horizontal')
+            ax_imag.set_xlabel('X (pixel)')
+            if iccd==0:
+                ax_imag.set_ylabel('Y (pixel)')
+
+            ax_imag.set_xlim(0, nx-1)
+            ax_imag.set_ylim(0, ny-1)
+            ax_imag.xaxis.set_major_locator(tck.MultipleLocator(500))
+            ax_imag.xaxis.set_minor_locator(tck.MultipleLocator(100))
+            ax_imag.yaxis.set_major_locator(tck.MultipleLocator(500))
+            ax_imag.yaxis.set_minor_locator(tck.MultipleLocator(100))
+            ax_imag.set_title('CCD {}'.format(iccd+1))
+            #ax_ycut.set_ylim(0, ny-1)
+
+            # plot histogram
+            bins = np.linspace(vmin, vmax, 50)
+            color = ('C0', 'C2', 'C3')[iccd]
+            ax_hist.hist(data.flatten(), bins=bins, color=color,
+                        label='CCD {}'.format(iccd+1))
+            ax_hist.legend(loc='upper right')
+            ax_hist.set_xticklabels([])
+            ax_hist.set_xlim(vmin, vmax)
+            ax_cbar.set_xlim(vmin, vmax)
+
+        self.suptitle('Bias')
+
+
+
+
 class TraceFigure(TraceFigureCommon):
     """Figure to plot the order tracing.
     """
     def __init__(self):
-        TraceFigureCommon.__init__(self, figsize=(20,10), dpi=150)
+        TraceFigureCommon.__init__(self, figsize=(10,6), dpi=300)
         self.ax1 = self.add_axes([0.05,0.07,0.50,0.86])
         self.ax2 = self.add_axes([0.59,0.55,0.36,0.34])
         self.ax3 = self.add_axes([0.59,0.13,0.36,0.34])
